Add files via upload
PSP_fastRadialScans.txt is a file listing the time ranges used for this project, when PSP is moving near-radially both toward and away from the Sun.

scaled.ipynb is the Jupyter Notebook file used to generate plots of magnetic field and velocity in each coordinate axis, mapping proton density, temperature, as well as calculating the speed, plasma beta, and mach number for each of these time series.

This does not yet include PSD or other data operations. Just a scaled version of the data prep phase.